+++ ext/tmdbsimple/account.py
@@ -3,7 +3,7 @@
 """
 tmdbsimple.account
 ~~~~~~~~~~~~~~~~~~
-This module implements the Account, Authentication, and Lists functionality 
+This module implements the Account, Authentication, and Lists functionality
 of tmdbsimple.
 
 Created by Celia Oakley on 2013-10-31.
@@ -25,7 +25,7 @@ class Account(TMDB):
     BASE_PATH = 'account'
     URLS = {
         'info': '',
-        'lists': '/{id}/lists', 
+        'lists': '/{id}/lists',
         'favorite_movies': '/{id}/favorite/movies',
         'favorite_tv': '/{id}/favorite/tv',
         'favorite': '/{id}/favorite',
@@ -43,9 +43,9 @@ class Account(TMDB):
 
     def info(self, **kwargs):
         """
-        Get the basic information for an account.
+        Get your account details.
 
-        Call this method first, before calling other Account methods.
+        Args:
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -57,14 +57,15 @@ class Account(TMDB):
         self.id = response['id']
         self._set_attrs_to_values(response)
         return response
-        
+
     def lists(self, **kwargs):
         """
-        Get the lists that you have created and marked as a favorite.
+        Get all of the lists created by an account. Will include private lists
+        if you are the owner.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
             language: (optional) ISO 639-1 code.
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -78,12 +79,12 @@ class Account(TMDB):
 
     def favorite_movies(self, **kwargs):
         """
-        Get the list of favorite movies for an account.
+        Get the list of your favorite movies.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
             language: (optional) ISO 639-1 code.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -97,12 +98,12 @@ class Account(TMDB):
 
     def favorite_tv(self, **kwargs):
         """
-        Get the list of favorite TV series for an account.
+        Get the list of your favorite TV shows.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
             language: (optional) ISO 639-1 code.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -116,7 +117,7 @@ class Account(TMDB):
 
     def favorite(self, **kwargs):
         """
-        Add or remove a movie to an accounts favorite list.
+        This method allows you to mark a movie or TV show as a favorite item.
 
         Args:
             media_type: 'movie' | 'tv'
@@ -130,8 +131,8 @@ class Account(TMDB):
         kwargs.update({'session_id': self.session_id})
 
         payload = {
-            'media_type': kwargs.pop('media_type', None), 
-            'media_id': kwargs.pop('media_id', None), 
+            'media_type': kwargs.pop('media_type', None),
+            'media_id': kwargs.pop('media_id', None),
             'favorite': kwargs.pop('favorite', None),
         }
 
@@ -141,12 +142,12 @@ class Account(TMDB):
 
     def rated_movies(self, **kwargs):
         """
-        Get the list of rated movies (and associated rating) for an account.
+        Get a list of all the movies you have rated.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
-            language: (optional) ISO 639-1 code.
+            language: (optional) ISO 639-1 value.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -160,12 +161,12 @@ class Account(TMDB):
 
     def rated_tv(self, **kwargs):
         """
-        Get the list of rated TV shows (and associated rating) for an account.
+        Get a list of all the TV shows you have rated.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
-            language: (optional) ISO 639-1 code.
+            language: (optional) ISO 639-1 value.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -182,9 +183,9 @@ class Account(TMDB):
         Get a list of all the TV episodes you have rated.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
-            language: (optional) ISO 639-1 code.
+            language: (optional) ISO 639-1 value.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -198,12 +199,12 @@ class Account(TMDB):
 
     def watchlist_movies(self, **kwargs):
         """
-        Get the list of movies on an account watchlist.
+        Get a list of all the movies you have added to your watchlist.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
-            language: (optional) ISO 639-1 code.
+            language: (optional) ISO 639-1 value.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -217,12 +218,12 @@ class Account(TMDB):
 
     def watchlist_tv(self, **kwargs):
         """
-        Get the list of TV series on an account watchlist.
+        Get a list of all the TV shows you have added to your watchlist.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
-            language: (optional) ISO 639-1 code.
+            language: (optional) ISO 639-1 value.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+            page: (optional) Minimum 1, maximum 1000, default 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -236,7 +237,7 @@ class Account(TMDB):
 
     def watchlist(self, **kwargs):
         """
-        Add or remove a movie to an accounts watch list.
+        Add a movie or TV show to your watchlist.
 
         Args:
             media_type: 'movie' | 'tv'
@@ -250,14 +251,15 @@ class Account(TMDB):
         kwargs.update({'session_id': self.session_id})
 
         payload = {
-            'media_type': kwargs.pop('media_type', None), 
-            'media_id': kwargs.pop('media_id', None), 
+            'media_type': kwargs.pop('media_type', None),
+            'media_id': kwargs.pop('media_id', None),
             'watchlist': kwargs.pop('watchlist', None),
         }
 
         response = self._POST(path, kwargs, payload)
         self._set_attrs_to_values(response)
         return response
+
 
 class Authentication(TMDB):
     """
@@ -268,24 +270,48 @@ class Authentication(TMDB):
     """
     BASE_PATH = 'authentication'
     URLS = {
+        'guest_session_new': '/guest_session/new',
         'token_new': '/token/new',
+        'session_new': '/session/new',
         'token_validate_with_login': '/token/validate_with_login',
-        'session_new': '/session/new', 
-        'guest_session_new': '/guest_session/new', 
+        'session_delete': '/session',
     }
+
+    def guest_session_new(self, **kwargs):
+        """
+        This method will let you create a new guest session. Guest sessions
+        are a type of session that will let a user rate movies and TV shows
+        but not require them to have a TMDb user account. More
+        information about user authentication can be found here
+        (https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id).
+
+        Please note, you should only generate a single guest session per
+        user (or device) as you will be able to attach the ratings to a
+        TMDb user account in the future. There is also IP limits in place
+        so you should always make sure it's the end user doing the guest
+        session actions.
+
+        If a guest session is not used for the first time within 24 hours,
+        it will be automatically deleted.
+
+        Args:
+
+        Returns:
+            A dict respresentation of the JSON returned from the API.
+        """
+        path = self._get_path('guest_session_new')
+
+        response = self._GET(path, kwargs)
+        self._set_attrs_to_values(response)
+        return response
 
     def token_new(self, **kwargs):
         """
-        Generate a valid request token for user based authentication.
+        Create a temporary request token that can be used to validate a TMDb
+        user login. More details about how this works can be found here
+        (https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id).
 
-        A request token is required to ask the user for permission to
-        access their account.
-
-        After obtaining the request_token, either:
-        (1) Direct your user to:
-                https://www.themoviedb.org/authenticate/REQUEST_TOKEN
-        or:
-        (2) Call token_validate_with_login() below.
+        Args:
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -296,30 +322,12 @@ class Authentication(TMDB):
         self._set_attrs_to_values(response)
         return response
 
-    def token_validate_with_login(self, **kwargs):
-        """
-        Authenticate a user with a TMDb username and password.  The user
-        must have a verified email address and be registered on TMDb.
-
-        Args:
-            request_token: The token you generated for the user to approve.
-            username: The user's username on TMDb.
-            password: The user's password on TMDb.
-
-        Returns:
-            A dict respresentation of the JSON returned from the API.
-        """
-        path = self._get_path('token_validate_with_login')
-
-        response = self._GET(path, kwargs)
-        self._set_attrs_to_values(response)
-        return response
-
     def session_new(self, **kwargs):
         """
-        Generate a session id for user based authentication.
-
-        A session id is required in order to use any of the write methods.
+        You can use this method to create a fully valid session ID once a user
+        has validated the request token. More information about how this works
+        can be found here
+        (https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id).
 
         Args:
             request_token: The token you generated for the user to approve.
@@ -335,18 +343,55 @@ class Authentication(TMDB):
         self._set_attrs_to_values(response)
         return response
 
-    def guest_session_new(self, **kwargs):
+    def token_validate_with_login(self, **kwargs):
         """
-        Generate a guest session id.
+        This method allows an application to validate a request token by entering
+        a username and password.
+
+        Not all applications have access to a web view so this can be used as a
+        substitute.
+
+        Please note, the preferred method of validating a request token is to
+        have a user authenticate the request via the TMDb website. You can read
+        about that method here
+        (https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id).
+
+        If you decide to use this method please use HTTPS.
+
+        Args:
+            username: The user's username on TMDb.
+            password: The user's password on TMDb.
+            request_token: The token you generated for the user to approve.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
-        path = self._get_path('guest_session_new')
+        path = self._get_path('token_validate_with_login')
 
         response = self._GET(path, kwargs)
         self._set_attrs_to_values(response)
         return response
+
+    def session_delete(self, **kwargs):
+        """
+        If you would like to delete (or "logout") from a session, call this
+        method with a valid session ID.
+
+        Args:
+
+        Returns:
+            A dict respresentation of the JSON returned from the API.
+        """
+        path = self._get_path('session_delete')
+
+        payload = {
+            'session_id': kwargs.pop('session_id', None),
+        }
+
+        response = self._DELETE(path, kwargs, payload)
+        self._set_attrs_to_values(response)
+        return response
+
 
 class GuestSessions(TMDB):
     """
@@ -356,7 +401,9 @@ class GuestSessions(TMDB):
     """
     BASE_PATH = 'guest_session'
     URLS = {
-        'rated_movies': '/{guest_session_id}/rated_movies',
+        'rated_movies': '/{guest_session_id}/rated/movies',
+        'rated_tv': '/{guest_session_id}/rated/tv',
+        'rated_tv_episodes': '/{guest_session_id}/rated/tv/episodes',
     }
 
     def __init__(self, guest_session_id=0):
@@ -365,12 +412,11 @@ class GuestSessions(TMDB):
 
     def rated_movies(self, **kwargs):
         """
-        Get a list of rated moview for a specific guest session id.
+        Get the rated movies for a guest session.
 
         Args:
-            page: (optional) Minimum 1, maximum 1000.
-            sort_by: (optional) 'created_at.asc' | 'created_at.desc'
             language: (optional) ISO 639-1 code.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -380,7 +426,41 @@ class GuestSessions(TMDB):
         response = self._GET(path, kwargs)
         self._set_attrs_to_values(response)
         return response
-    
+
+    def rated_tv(self, **kwargs):
+        """
+        Get the rated TV shows for a guest session.
+
+        Args:
+            language: (optional) ISO 639-1 code.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+
+        Returns:
+            A dict respresentation of the JSON returned from the API.
+        """
+        path = self._get_guest_session_id_path('rated_tv')
+
+        response = self._GET(path, kwargs)
+        self._set_attrs_to_values(response)
+        return response
+
+    def rated_tv_episodes(self, **kwargs):
+        """
+        Get the rated TV episodes for a guest session.
+
+        Args:
+            language: (optional) ISO 639-1 code.
+            sort_by: (optional) Allowed Values: created_at.asc, created_at.desc
+
+        Returns:
+            A dict respresentation of the JSON returned from the API.
+        """
+        path = self._get_guest_session_id_path('rated_tv_episodes')
+
+        response = self._GET(path, kwargs)
+        self._set_attrs_to_values(response)
+        return response
+
 
 class Lists(TMDB):
     """
@@ -391,11 +471,12 @@ class Lists(TMDB):
     BASE_PATH = 'list'
     URLS = {
         'info': '/{id}',
-        'item_status': '/{id}/item_status', 
-        'create_list': '',
+        'item_status': '/{id}/item_status',
+        'list_create': '',
         'add_item': '/{id}/add_item',
         'remove_item': '/{id}/remove_item',
-        'clear': '/{id}/clear',
+        'list_clear': '/{id}/clear',
+        'list_delete': '/{id}',
     }
 
     def __init__(self, id=0, session_id=0):
@@ -405,7 +486,10 @@ class Lists(TMDB):
 
     def info(self, **kwargs):
         """
-        Get a list by id.
+        Get the details of a list.
+
+        Args:
+            language: (optional) ISO 639-1 code.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -418,10 +502,11 @@ class Lists(TMDB):
 
     def item_status(self, **kwargs):
         """
-        Check to see if a movie id is already added to a list.
+        You can use this method to check if a movie has already been added to
+        the list.
 
         Args:
-            movie_id: The id of the movie.
+            movie_id: The id of the movie.  Minimum 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -432,11 +517,9 @@ class Lists(TMDB):
         self._set_attrs_to_values(response)
         return response
 
-    def create_list(self, **kwargs):
+    def list_create(self, **kwargs):
         """
-        Create a new list.
-
-        A valid session id is required.
+        Create a list.
 
         Args:
             name: Name of the list.
@@ -446,28 +529,26 @@ class Lists(TMDB):
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
-        path = self._get_path('create_list')
+        path = self._get_path('list_create')
         kwargs.update({'session_id': self.session_id})
 
         payload = {
-            'name': kwargs.pop('name', None), 
+            'name': kwargs.pop('name', None),
             'description': kwargs.pop('description', None),
+            'language': kwargs.pop('language', None),
         }
-        if 'language' in kwargs:
-            payload['language'] = kwargs['language']
 
         response = self._POST(path, kwargs, payload)
         self._set_attrs_to_values(response)
+        self.id = self.list_id
         return response
 
     def add_item(self, **kwargs):
         """
-        Add new movies to a list that the user created.
-
-        A valid session id is required.
+        Add a movie to a list.
 
         Args:
-            media_id: A movie id.
+            media_id: A movie id.  Minimum 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -476,21 +557,19 @@ class Lists(TMDB):
         kwargs.update({'session_id': self.session_id})
 
         payload = {
-            'media_id': kwargs.pop('media_id', None), 
+            'media_id': kwargs.pop('media_id', None),
         }
 
         response = self._POST(path, kwargs, payload)
         self._set_attrs_to_values(response)
         return response
-        
+
     def remove_item(self, **kwargs):
         """
-        Delete movies from a list that the user created.
-
-        A valid session id is required.
+        Remove a movie from a list.
 
         Args:
-            media_id: A movie id.
+            media_id: A movie id.  Minimum 1.
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -499,19 +578,16 @@ class Lists(TMDB):
         kwargs.update({'session_id': self.session_id})
 
         payload = {
-            'media_id': kwargs.pop('media_id', None), 
+            'media_id': kwargs.pop('media_id', None),
         }
 
         response = self._POST(path, kwargs, payload)
         self._set_attrs_to_values(response)
         return response
 
-    def clear_list(self, **kwargs):
+    def list_clear(self, **kwargs):
         """
-        Clears all of the items within a list. This is an irreversible action
-        and should be treated with caution.
-
-        A valid session id is required.
+        Clear all of the items from a list.
 
         Args:
             confirm: True (do it) | False (don't do it)
@@ -519,11 +595,28 @@ class Lists(TMDB):
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
-        path = self._get_id_path('clear')
+        path = self._get_id_path('list_clear')
         kwargs.update({'session_id': self.session_id})
 
         payload = {}
 
         response = self._POST(path, kwargs, payload)
+        self._set_attrs_to_values(response)
+        return response
+
+    def list_delete(self, **kwargs):
+        """
+        Delete a list.
+
+        Args:
+            None
+
+        Returns:
+            A dict respresentation of the JSON returned from the API.
+        """
+        path = self._get_id_path('list_delete')
+        kwargs.update({'session_id': self.session_id})
+
+        response = self._DELETE(path, kwargs)
         self._set_attrs_to_values(response)
         return response