+++ ext/tmdbsimple/base.py
@@ -12,7 +12,6 @@ Created by Celia Oakley on 2013-10-31.
 """
 
 import json
-import requests
 
 
 class APIKeyError(Exception):
@@ -27,9 +26,10 @@ class TMDB(object):
     URLS = {}
 
     def __init__(self):
-        from . import API_VERSION
+        from . import API_VERSION, REQUESTS_SESSION
         self.base_uri = 'https://api.themoviedb.org'
         self.base_uri += '/{version}'.format(version=API_VERSION)
+        self.session = REQUESTS_SESSION
 
     def _get_path(self, key):
         return self.BASE_PATH + self.URLS[key]
@@ -40,17 +40,21 @@ class TMDB(object):
     def _get_guest_session_id_path(self, key):
         return self._get_path(key).format(
             guest_session_id=self.guest_session_id)
-    
+
     def _get_credit_id_path(self, key):
         return self._get_path(key).format(credit_id=self.credit_id)
 
+    def _get_media_type_time_window_path(self, key):
+        return self._get_path(key).format(
+            media_type=self.media_type, time_window=self.time_window)
+
     def _get_tv_id_season_number_path(self, key):
-        return self._get_path(key).format(tv_id=self.tv_id,
-            season_number=self.season_number)
+        return self._get_path(key).format(
+            tv_id=self.tv_id, season_number=self.season_number)
 
     def _get_tv_id_season_number_episode_number_path(self, key):
-        return self._get_path(key).format(tv_id=self.tv_id,
-            season_number=self.season_number,
+        return self._get_path(key).format(
+            tv_id=self.tv_id, season_number=self.season_number,
             episode_number=self.episode_number)
 
     def _get_complete_url(self, path):
@@ -76,8 +80,8 @@ class TMDB(object):
         url = self._get_complete_url(path)
         params = self._get_params(params)
 
-        response = requests.request(
-            method, url, params=params, 
+        response = self.session.request(
+            method, url, params=params,
             data=json.dumps(payload) if payload else payload,
             headers=self.headers)
 
@@ -108,4 +112,3 @@ class TMDB(object):
             for key in response.keys():
                 if not hasattr(self, key) or not callable(getattr(self, key)):
                     setattr(self, key, response[key])
-