+++ medusa/app.py
@@ -172,7 +172,11 @@ class MedusaApp(object):
 
         self.HANDLE_REVERSE_PROXY = False
         self.PROXY_SETTING = None
+        self.PROXY_PROVIDERS = True
         self.PROXY_INDEXERS = True
+        self.PROXY_CLIENTS = True
+        self.PROXY_OTHERS = True
+
         self.SSL_VERIFY = True
         self.SSL_CA_BUNDLE = None
 
@@ -226,6 +230,8 @@ class MedusaApp(object):
         self.INDEXER_TIMEOUT = None
         self.SCENE_DEFAULT = False
         self.ANIME_DEFAULT = False
+        self.SHOWLISTS_DEFAULT = ['series']
+        self.SHOWLISTS_DEFAULT_ANIME = []
         self.PROVIDER_ORDER = []
 
         self.NAMING_MULTI_EP = False