+++ medusa/tv/episode.py
@@ -76,7 +76,6 @@ from medusa.sbdatetime import sbdatetime
 from medusa.scene_numbering import (
     get_scene_absolute_numbering,
     get_scene_numbering,
-    xem_refresh,
 )
 from medusa.tv.base import Identifier, TV
 
@@ -260,9 +259,9 @@ class Episode(TV):
         self.version = 0
         self.release_group = ''
         self._location = filepath
-        self.scene_season = 0
-        self.scene_episode = 0
-        self.scene_absolute_number = 0
+        self._scene_season = None
+        self._scene_episode = 0
+        self._scene_absolute_number = 0
         self.manually_searched = False
         self.related_episodes = []
         self.wanted_quality = []
@@ -425,6 +424,45 @@ class Episode(TV):
     def status_name(self):
         """Return the status name."""
         return statusStrings[self.status]
+
+    @property
+    def scene_season(self):
+        """Return the scene season."""
+        if self._scene_season is None:
+            return self.season
+
+        return self._scene_season
+
+    @scene_season.setter
+    def scene_season(self, value):
+        """Set the scene season."""
+        self._scene_season = try_int(value, None)
+
+    @property
+    def scene_episode(self):
+        """Return the scene episode."""
+        if not self._scene_episode:
+            return self.episode
+
+        return self._scene_episode
+
+    @scene_episode.setter
+    def scene_episode(self, value):
+        """Set the scene episode."""
+        self._scene_episode = try_int(value, 0)
+
+    @property
+    def scene_absolute_number(self):
+        """Return the scene absolute number."""
+        if not self._scene_absolute_number:
+            return self.absolute_number
+
+        return self._scene_absolute_number
+
+    @scene_absolute_number.setter
+    def scene_absolute_number(self, value):
+        """Set the scene absolute number."""
+        self._scene_absolute_number = try_int(value, 0)
 
     @property
     def quality_name(self):
@@ -664,12 +702,9 @@ class Episode(TV):
             self.indexerid = int(sql_results[0]['indexerid'])
             self.indexer = int(sql_results[0]['indexer'])
 
-            # FIXME: This shouldn't be part of a possible apiv2 episodes request
-            xem_refresh(self.series)
-
-            self.scene_season = try_int(sql_results[0]['scene_season'], 0)
-            self.scene_episode = try_int(sql_results[0]['scene_episode'], 0)
-            self.scene_absolute_number = try_int(sql_results[0]['scene_absolute_number'], 0)
+            self.scene_season = sql_results[0]['scene_season']
+            self.scene_episode = sql_results[0]['scene_episode']
+            self.scene_absolute_number = sql_results[0]['scene_absolute_number']
 
             if self.scene_absolute_number == 0:
                 self.scene_absolute_number = get_scene_absolute_numbering(
@@ -677,10 +712,9 @@ class Episode(TV):
                     self.absolute_number
                 )
 
-            if self.scene_season == 0 or self.scene_episode == 0:
+            if self.scene_season is None or self.scene_episode == 0:
                 self.scene_season, self.scene_episode = get_scene_numbering(
-                    self.series,
-                    self.season, self.episode
+                    self.series, self.episode, self.season
                 )
 
             if sql_results[0]['release_name'] is not None:
@@ -825,16 +859,15 @@ class Episode(TV):
         self.season = season
         self.episode = episode
 
-        xem_refresh(self.series)
-
         self.scene_absolute_number = get_scene_absolute_numbering(
             self.series,
             self.absolute_number
         )
 
+        # TODO: Just me not understanding. If we're getting the show info from the indexer.
+        # Why are we trying to get the scene_season and scene_episode from the db?
         self.scene_season, self.scene_episode = get_scene_numbering(
-            self.series,
-            self.season, self.episode
+            self.series, self.episode, self.season
         )
 
         self.description = getattr(my_ep, 'overview', '')
@@ -1027,9 +1060,7 @@ class Episode(TV):
                     )
 
                     self.scene_season, self.scene_episode = get_scene_numbering(
-                        self.series.series_id,
-                        self.series.indexer,
-                        self.season, self.episode
+                        self.series, self.episode, self.season
                     )
 
                     self.description = ep_details.findtext('plot')
@@ -1560,9 +1591,9 @@ class Episode(TV):
             '%E': str(self.episode),
             '%0E': '%02d' % self.episode,
             '%XS': str(self.scene_season),
-            '%0XS': '%02d' % self.scene_season,
+            '%0XS': '%02d' % try_int(self.scene_season, self.season),
             '%XE': str(self.scene_episode),
-            '%0XE': '%02d' % self.scene_episode,
+            '%0XE': '%02d' % try_int(self.scene_episode, self.episode),
             '%AB': '%(#)03d' % {'#': self.absolute_number},
             '%XAB': '%(#)03d' % {'#': self.scene_absolute_number},
             '%RN': release_name(self.release_name),