+++ themes-default/slim/src/components/edit-show.vue
@@ -86,7 +86,8 @@
                                 </config-toggle-slider>
 
                                 <config-toggle-slider :value="show.config.anime" @input="changeFormat($event, 'anime')" label="Anime" id="anime">
-                                    <span>enable if the show is Anime and episodes are released as Show.265 rather than Show.S02E03</span>
+                                    <span>enable if the shows episodes are released using absolute numbering</span>
+                                    <p>For example as Show.265 rather than Show.S02E03</p>
                                 </config-toggle-slider>
 
                                 <config-template v-if="show.config.anime" label-for="anidbReleaseGroup" label="Release Groups">
@@ -281,6 +282,7 @@ export default {
     },
     computed: {
         ...mapState({
+            general: state => state.config.general,
             indexers: state => state.config.indexers,
             anime: state => state.config.anime,
             layout: state => state.config.layout,
@@ -465,7 +467,7 @@ export default {
             this.show.language = value;
         },
         changeFormat(value, formatOption) {
-            const { anime } = this;
+            const { anime, general } = this;
             this.show.config[formatOption] = value;
             if (value) {
                 // Check each format option, disable the other options.
@@ -477,12 +479,12 @@ export default {
             if (formatOption === 'anime' && anime.autoAnimeToList) {
                 if (value) {
                     // Auto anime to list is enabled. If changing the show format to anime, add 'Anime' to show lists.
-                    this.showLists.push('anime');
+                    this.showLists = anime.showlistDefaultAnime;
                     // The filter makes sure there are unique strings.
                     this.showLists = this.showLists.filter((v, i, a) => a.indexOf(v) === i);
                 } else {
-                    // Auto anime to list is enabled. If changing the show format to anime, add 'Anime' to show lists.
-                    this.showLists = this.showLists.filter(list => list !== 'anime');
+                    // Return to default show lists.
+                    this.showLists = general.showDefaults.showLists;
                 }
             }
         }