+++ themes-default/slim/src/components/show-results.vue
@@ -31,8 +31,9 @@
                     <template slot="table-row" slot-scope="props">
                         <span v-if="props.column.label === 'Provider'" class="align-center">
                             <img :src="`images/providers/${props.row.provider.imageName}`"
-                                 :alt="props.row.provider.name" width="16" class="addQTip"
+                                 :alt="props.row.provider.name" width="16"
                                  :title="props.row.provider.name"
+                                 v-tooltip.right="props.row.provider.name"
                                  onError="this.onerror=null;this.src='images/providers/missing.png';"
                             >
                         </span>
@@ -70,6 +71,10 @@
                         </span>
                     </template>
 
+                    <div slot="emptystate">
+                        No search results available
+                    </div>
+
                 </vue-good-table>
             </div>
         </div>
@@ -84,7 +89,7 @@ import { manageCookieMixin } from '../mixins/manage-cookie';
 import { StateSwitch } from './helpers';
 import QualityPill from './helpers/quality-pill.vue';
 import { episodeToSlug, humanFileSize } from '../utils/core';
-import { addQTip } from '../utils/jquery';
+import { VTooltip } from 'v-tooltip';
 
 export default {
     name: 'show-results',
@@ -92,6 +97,9 @@ export default {
         VueGoodTable,
         StateSwitch,
         QualityPill
+    },
+    directives: {
+        tooltip: VTooltip
     },
     mixins: [
         manageCookieMixin('showResults')
@@ -204,8 +212,6 @@ export default {
         if (result.providersSearched > 0 && result.totalSearchResults.length === 0) {
             forceSearch();
         }
-
-        addQTip();
     },
     computed: {
         ...mapState({
@@ -396,9 +402,6 @@ export default {
                     await getShowEpisodeHistory({ showSlug: show.id.slug, episodeSlug });
                     await getProviderCacheResults({ showSlug: show.id.slug, season, episode });
                 }
-
-                // Update qTip
-                addQTip();
             },
             deep: true,
             immediate: false
@@ -407,4 +410,5 @@ export default {
 };
 </script>
 <style scoped>
+@import '../style/v-tooltip.css';
 </style>